+++ EpicLoot/src/Crafting/EnchantCostsConfig.cs
@@ -57,8 +57,9 @@ namespace EpicLoot.Crafting
     [Serializable]
     public class IdentifyCostConfig
     {
-        public ItemRarity Rarity;
+        public Heightmap.Biome Biome;
         public Dictionary<LootRoller.LootRollCategories, List<ItemAmountConfig>> CostByIDType = new Dictionary<LootRoller.LootRollCategories, List<ItemAmountConfig>>();
+        public Dictionary<ItemRarity, List<ItemAmountConfig>> CostByRarity = new Dictionary<ItemRarity, List<ItemAmountConfig>>();
     }
 
     [Serializable]
@@ -71,6 +72,6 @@ namespace EpicLoot.Crafting
         public List<ItemAmountConfig> ReAugmentCosts = new List<ItemAmountConfig>();
         public List<RuneCostConfig> RuneExtractCosts = new List<RuneCostConfig>();
         public List<RuneCostConfig> RuneEtchCosts = new List<RuneCostConfig>();
-        public List<IdentifyCostConfig> IdentifyCosts = new List<IdentifyCostConfig>();
+        public Dictionary<Heightmap.Biome, IdentifyCostConfig> IdentifyCosts = new Dictionary<Heightmap.Biome, IdentifyCostConfig>();
     }
 }